requirements.txt
amqp==5.0.6
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.1
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==6.2
Deprecated==1.2.13
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
flower==1.0.0
gevent==21.8.0
greenlet==1.1.2
humanize==3.12.0
importlib-metadata==4.8.2
iniconfig==1.1.1
kombu==5.2.2
packaging==21.3
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
psycopg2==2.9.2
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
python-dotenv==0.19.2
pytz==2021.3
redis==4.0.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
tornado==6.1
typing_extensions==4.0.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
